Split out payload and state from track and trace handler
This reduces the size of the handler files as well as
splitting out unrelated components.
