+++ contracts/track_and_trace/src/handler.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 use protobuf;
-use protobuf::Message;
 use protobuf::RepeatedField;
 
 use std::collections::HashMap;
@@ -34,554 +33,25 @@ cfg_if! {
 }
 
 use grid_sdk::protos::track_and_trace_agent::TrackAndTraceAgent as Agent;
-use grid_sdk::protos::track_and_trace_agent::TrackAndTraceAgentContainer as AgentContainer;
 use grid_sdk::protos::track_and_trace_payload::CreateTrackAndTraceAgentAction as CreateAgentAction;
 use grid_sdk::protos::track_and_trace_payload::{
     AnswerProposalAction, AnswerProposalAction_Response, CreateProposalAction, CreateRecordAction,
-    CreateRecordTypeAction, FinalizeRecordAction, RevokeReporterAction, SCPayload,
-    SCPayload_Action, UpdatePropertiesAction,
+    CreateRecordTypeAction, FinalizeRecordAction, RevokeReporterAction, UpdatePropertiesAction,
 };
 use grid_sdk::protos::track_and_trace_property::{
-    Property, PropertyContainer, PropertyPage, PropertyPageContainer, PropertyPage_ReportedValue,
-    PropertySchema, PropertySchema_DataType, Property_Reporter, TrackAndTracePropertyValue,
+    Property, PropertyPage, PropertyPage_ReportedValue, PropertySchema, PropertySchema_DataType,
+    Property_Reporter, TrackAndTracePropertyValue,
 };
 use grid_sdk::protos::track_and_trace_proposal::{
     Proposal, ProposalContainer, Proposal_Role, Proposal_Status,
 };
-use grid_sdk::protos::track_and_trace_record::{
-    Record, RecordContainer, RecordType, RecordTypeContainer, Record_AssociatedAgent,
-};
+use grid_sdk::protos::track_and_trace_record::{Record, RecordType, Record_AssociatedAgent};
 
 use crate::addressing::*;
+use crate::payload::{Action, SupplyChainPayload};
+use crate::state::SupplyChainState;
 
 const PROPERTY_PAGE_MAX_LENGTH: usize = 256;
-
-#[derive(Debug, Clone)]
-enum Action {
-    CreateAgent(CreateAgentAction),
-    CreateRecord(CreateRecordAction),
-    FinalizeRecord(FinalizeRecordAction),
-    CreateRecordType(CreateRecordTypeAction),
-    UpdateProperties(UpdatePropertiesAction),
-    CreateProposal(CreateProposalAction),
-    AnswerProposal(AnswerProposalAction),
-    RevokeReporter(RevokeReporterAction),
-}
-
-struct SupplyChainPayload {
-    action: Action,
-    timestamp: u64,
-}
-
-impl SupplyChainPayload {
-    pub fn new(payload: &[u8]) -> Result<Option<SupplyChainPayload>, ApplyError> {
-        let payload: SCPayload = match protobuf::parse_from_bytes(payload) {
-            Ok(payload) => payload,
-            Err(_) => {
-                return Err(ApplyError::InvalidTransaction(String::from(
-                    "Cannot deserialize payload",
-                )));
-            }
-        };
-
-        let supply_chain_action = payload.get_action();
-        let action = match supply_chain_action {
-            SCPayload_Action::CREATE_AGENT => {
-                let create_agent = payload.get_create_agent();
-                if create_agent.get_name() == "" {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Agent name cannot be an empty string",
-                    )));
-                }
-                Action::CreateAgent(create_agent.clone())
-            }
-            SCPayload_Action::CREATE_RECORD => {
-                let create_record = payload.get_create_record();
-                if create_record.get_record_id() == "" {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Record id cannot be empty string",
-                    )));
-                }
-                Action::CreateRecord(create_record.clone())
-            }
-            SCPayload_Action::FINALIZE_RECORD => {
-                Action::FinalizeRecord(payload.get_finalize_record().clone())
-            }
-            SCPayload_Action::CREATE_RECORD_TYPE => {
-                let create_record_type = payload.get_create_record_type();
-                if create_record_type.get_name() == "" {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Record Type name cannot be an empty string",
-                    )));
-                };
-                let properties = create_record_type.get_properties();
-                if properties.len() == 0 {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Record type must have at least one property",
-                    )));
-                }
-                for prop in properties {
-                    if prop.name == "" {
-                        return Err(ApplyError::InvalidTransaction(String::from(
-                            "Property name cannot be an empty string",
-                        )));
-                    }
-                }
-
-                Action::CreateRecordType(create_record_type.clone())
-            }
-            SCPayload_Action::UPDATE_PROPERTIES => {
-                Action::UpdateProperties(payload.get_update_properties().clone())
-            }
-            SCPayload_Action::CREATE_PROPOSAL => {
-                Action::CreateProposal(payload.get_create_proposal().clone())
-            }
-            SCPayload_Action::ANSWER_PROPOSAL => {
-                Action::AnswerProposal(payload.get_answer_proposal().clone())
-            }
-            SCPayload_Action::REVOKE_REPORTER => {
-                Action::RevokeReporter(payload.get_revoke_reporter().clone())
-            }
-        };
-        let timestamp = match payload.get_timestamp() {
-            0 => {
-                return Err(ApplyError::InvalidTransaction(String::from(
-                    "Timestamp is not set",
-                )));
-            }
-            x => x,
-        };
-
-        Ok(Some(SupplyChainPayload {
-            action: action,
-            timestamp: timestamp,
-        }))
-    }
-
-    pub fn get_action(&self) -> Action {
-        self.action.clone()
-    }
-
-    pub fn get_timestamp(&self) -> u64 {
-        self.timestamp
-    }
-}
-
-pub struct SupplyChainState<'a> {
-    context: &'a mut TransactionContext,
-}
-
-impl<'a> SupplyChainState<'a> {
-    pub fn new(context: &'a mut TransactionContext) -> SupplyChainState {
-        SupplyChainState { context: context }
-    }
-
-    pub fn get_record(&mut self, record_id: &str) -> Result<Option<Record>, ApplyError> {
-        let address = make_record_address(record_id);
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let records: RecordContainer = match protobuf::parse_from_bytes(packed.as_slice()) {
-                    Ok(records) => records,
-                    Err(_) => {
-                        return Err(ApplyError::InternalError(String::from(
-                            "Cannot deserialize record container",
-                        )));
-                    }
-                };
-
-                for record in records.get_entries() {
-                    if record.record_id == record_id {
-                        return Ok(Some(record.clone()));
-                    }
-                }
-                Ok(None)
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn set_record(&mut self, record_id: &str, record: Record) -> Result<(), ApplyError> {
-        let address = make_record_address(record_id);
-        let d = self.context.get_state_entry(&address)?;
-        let mut record_container = match d {
-            Some(packed) => match protobuf::parse_from_bytes(packed.as_slice()) {
-                Ok(records) => records,
-                Err(_) => {
-                    return Err(ApplyError::InternalError(String::from(
-                        "Cannot deserialize record container",
-                    )));
-                }
-            },
-            None => RecordContainer::new(),
-        };
-        // remove old record if it exists and sort the records by record id
-        let records = record_container.get_entries().to_vec();
-        let mut index = None;
-        let mut count = 0;
-        for record in records.clone() {
-            if record.record_id == record_id {
-                index = Some(count);
-                break;
-            }
-            count = count + 1;
-        }
-
-        match index {
-            Some(x) => {
-                record_container.entries.remove(x);
-            }
-            None => (),
-        };
-        record_container.entries.push(record);
-        record_container
-            .entries
-            .sort_by_key(|r| r.clone().record_id);
-        let serialized = match record_container.write_to_bytes() {
-            Ok(serialized) => serialized,
-            Err(_) => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot serialize record container",
-                )));
-            }
-        };
-        self.context
-            .set_state_entry(address, serialized)
-            .map_err(|err| ApplyError::InternalError(format!("{}", err)))?;
-        Ok(())
-    }
-
-    pub fn get_record_type(&mut self, type_name: &str) -> Result<Option<RecordType>, ApplyError> {
-        let address = make_record_type_address(type_name);
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let record_types: RecordTypeContainer =
-                    match protobuf::parse_from_bytes(packed.as_slice()) {
-                        Ok(record_types) => record_types,
-                        Err(_) => {
-                            return Err(ApplyError::InternalError(String::from(
-                                "Cannot deserialize record type container",
-                            )));
-                        }
-                    };
-
-                for record_type in record_types.get_entries() {
-                    if record_type.name == type_name {
-                        return Ok(Some(record_type.clone()));
-                    }
-                }
-                Ok(None)
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn set_record_type(
-        &mut self,
-        type_name: &str,
-        record_type: RecordType,
-    ) -> Result<(), ApplyError> {
-        let address = make_record_type_address(type_name);
-        let d = self.context.get_state_entry(&address)?;
-        let mut record_types = match d {
-            Some(packed) => match protobuf::parse_from_bytes(packed.as_slice()) {
-                Ok(record_types) => record_types,
-                Err(_) => {
-                    return Err(ApplyError::InternalError(String::from(
-                        "Cannot deserialize record container",
-                    )));
-                }
-            },
-            None => RecordTypeContainer::new(),
-        };
-
-        record_types.entries.push(record_type);
-        record_types.entries.sort_by_key(|rt| rt.clone().name);
-        let serialized = match record_types.write_to_bytes() {
-            Ok(serialized) => serialized,
-            Err(_) => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot serialize record type container",
-                )));
-            }
-        };
-        self.context
-            .set_state_entry(address, serialized)
-            .map_err(|err| ApplyError::InternalError(format!("{}", err)))?;
-        Ok(())
-    }
-
-    pub fn get_agent(&mut self, agent_id: &str) -> Result<Option<Agent>, ApplyError> {
-        let address = make_agent_address(agent_id);
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let agents: AgentContainer = match protobuf::parse_from_bytes(packed.as_slice()) {
-                    Ok(agents) => agents,
-                    Err(_) => {
-                        return Err(ApplyError::InternalError(String::from(
-                            "Cannot deserialize agent container",
-                        )));
-                    }
-                };
-
-                for agent in agents.get_entries() {
-                    if agent.public_key == agent_id {
-                        return Ok(Some(agent.clone()));
-                    }
-                }
-                Ok(None)
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn set_agent(&mut self, agent_id: &str, agent: Agent) -> Result<(), ApplyError> {
-        let address = make_agent_address(agent_id);
-        let d = self.context.get_state_entry(&address)?;
-        let mut agents = match d {
-            Some(packed) => match protobuf::parse_from_bytes(packed.as_slice()) {
-                Ok(agents) => agents,
-                Err(_) => {
-                    return Err(ApplyError::InternalError(String::from(
-                        "Cannot deserialize agent container",
-                    )));
-                }
-            },
-            None => AgentContainer::new(),
-        };
-
-        agents.entries.push(agent);
-        agents.entries.sort_by_key(|a| a.clone().public_key);
-        let serialized = match agents.write_to_bytes() {
-            Ok(serialized) => serialized,
-            Err(_) => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot serialize agent container",
-                )));
-            }
-        };
-        self.context
-            .set_state_entry(address, serialized)
-            .map_err(|err| ApplyError::InternalError(format!("{}", err)))?;
-        Ok(())
-    }
-
-    pub fn get_property(
-        &mut self,
-        record_id: &str,
-        property_name: &str,
-    ) -> Result<Option<Property>, ApplyError> {
-        let address = make_property_address(record_id, property_name, 0);
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let properties: PropertyContainer =
-                    match protobuf::parse_from_bytes(packed.as_slice()) {
-                        Ok(properties) => properties,
-                        Err(_) => {
-                            return Err(ApplyError::InternalError(String::from(
-                                "Cannot deserialize property container",
-                            )));
-                        }
-                    };
-
-                for property in properties.get_entries() {
-                    if property.name == property_name {
-                        return Ok(Some(property.clone()));
-                    }
-                }
-                Ok(None)
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn set_property(
-        &mut self,
-        record_id: &str,
-        property_name: &str,
-        property: Property,
-    ) -> Result<(), ApplyError> {
-        let address = make_property_address(record_id, property_name, 0);
-        let d = self.context.get_state_entry(&address)?;
-        let mut property_container = match d {
-            Some(packed) => match protobuf::parse_from_bytes(packed.as_slice()) {
-                Ok(properties) => properties,
-                Err(_) => {
-                    return Err(ApplyError::InternalError(String::from(
-                        "Cannot deserialize property container",
-                    )));
-                }
-            },
-            None => PropertyContainer::new(),
-        };
-        // remove old property if it exists and sort the properties by name
-        let properties = property_container.get_entries().to_vec();
-        let mut index = None;
-        let mut count = 0;
-        for prop in properties.clone() {
-            if prop.name == property_name {
-                index = Some(count);
-                break;
-            }
-            count = count + 1;
-        }
-
-        match index {
-            Some(x) => {
-                property_container.entries.remove(x);
-            }
-            None => (),
-        };
-        property_container.entries.push(property);
-        property_container.entries.sort_by_key(|p| p.clone().name);
-        let serialized = match property_container.write_to_bytes() {
-            Ok(serialized) => serialized,
-            Err(_) => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot serialize property container",
-                )));
-            }
-        };
-        self.context
-            .set_state_entry(address, serialized)
-            .map_err(|err| ApplyError::InternalError(format!("{}", err)))?;
-        Ok(())
-    }
-
-    pub fn get_property_page(
-        &mut self,
-        record_id: &str,
-        property_name: &str,
-        page: u32,
-    ) -> Result<Option<PropertyPage>, ApplyError> {
-        let address = make_property_address(record_id, property_name, page);
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let property_pages: PropertyPageContainer =
-                    match protobuf::parse_from_bytes(packed.as_slice()) {
-                        Ok(property_pages) => property_pages,
-                        Err(_) => {
-                            return Err(ApplyError::InternalError(String::from(
-                                "Cannot deserialize property page container",
-                            )));
-                        }
-                    };
-
-                for property_page in property_pages.get_entries() {
-                    if property_page.name == property_name {
-                        return Ok(Some(property_page.clone()));
-                    }
-                }
-                Ok(None)
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn set_property_page(
-        &mut self,
-        record_id: &str,
-        property_name: &str,
-        page_num: u32,
-        property_page: PropertyPage,
-    ) -> Result<(), ApplyError> {
-        let address = make_property_address(record_id, property_name, page_num);
-        let d = self.context.get_state_entry(&address)?;
-        let mut property_pages = match d {
-            Some(packed) => match protobuf::parse_from_bytes(packed.as_slice()) {
-                Ok(property_pages) => property_pages,
-                Err(_) => {
-                    return Err(ApplyError::InternalError(String::from(
-                        "Cannot deserialize property page container",
-                    )));
-                }
-            },
-            None => PropertyPageContainer::new(),
-        };
-        // remove old property page if it exists and sort the property pages by name
-        let pages = property_pages.get_entries().to_vec();
-        let mut index = None;
-        let mut count = 0;
-        for page in pages.clone() {
-            if page.name == property_name {
-                index = Some(count);
-                break;
-            }
-            count = count + 1;
-        }
-
-        match index {
-            Some(x) => {
-                property_pages.entries.remove(x);
-            }
-            None => (),
-        };
-        property_pages.entries.push(property_page);
-        property_pages.entries.sort_by_key(|pp| pp.clone().name);
-        let serialized = match property_pages.write_to_bytes() {
-            Ok(serialized) => serialized,
-            Err(_) => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot serialize property page container",
-                )));
-            }
-        };
-        self.context
-            .set_state_entry(address, serialized)
-            .map_err(|err| ApplyError::InternalError(format!("{}", err)))?;
-        Ok(())
-    }
-
-    pub fn get_proposal_container(
-        &mut self,
-        record_id: &str,
-        agent_id: &str,
-    ) -> Result<Option<ProposalContainer>, ApplyError> {
-        let address = make_proposal_address(record_id, agent_id);
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let proposals: ProposalContainer =
-                    match protobuf::parse_from_bytes(packed.as_slice()) {
-                        Ok(property_pages) => property_pages,
-                        Err(_) => {
-                            return Err(ApplyError::InternalError(String::from(
-                                "Cannot deserialize proposal container",
-                            )));
-                        }
-                    };
-
-                Ok(Some(proposals))
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn set_proposal_container(
-        &mut self,
-        record_id: &str,
-        agent_id: &str,
-        proposals: ProposalContainer,
-    ) -> Result<(), ApplyError> {
-        let address = make_proposal_address(record_id, agent_id);
-        let serialized = match proposals.write_to_bytes() {
-            Ok(serialized) => serialized,
-            Err(_) => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot serialize proposal container",
-                )));
-            }
-        };
-        self.context
-            .set_state_entry(address, serialized)
-            .map_err(|err| ApplyError::InternalError(format!("{}", err)))?;
-        Ok(())
-    }
-}
 
 pub struct SupplyChainTransactionHandler {
     family_name: String,
